backend/requirements.txt
# FastAPI Backend Requirements for PowerBI Web Replica

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Data Processing
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0
pyarrow>=14.0.0,<15.0.0

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Object Storage (S3/MinIO)
boto3==1.34.0
minio==7.2.0

# Background Jobs
celery==5.3.4
kombu==5.3.4

# Visualization & Export
altair==5.2.0
matplotlib==3.8.2
reportlab==4.0.7
Pillow==10.1.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Configuration
python-dotenv==1.0.0